Update fix-html-snapshot.js
When test fails it always gets snapshot.source as undefined, leading to displaying this error of replace is not defined in undefined rather then showing initial cause of error in failed test. probably we need to make sure snapshot always have source, even empty one returned when it's created, but it's easier fix, considering i'm not aware of whole codebase yet.